Add json convenience methods to websocket message objects
