+++ docs/web_reference.rst
@@ -260,9 +260,9 @@ like one using :meth:`Request.copy`.
 
          async def json(self, *, loads=json.loads):
              body = await self.text()
-             return loader(body)
+             return loads(body)
 
-      :param callable loader: any :term:`callable` that accepts
+      :param callable loads: any :term:`callable` that accepts
                               :class:`str` and returns :class:`dict`
                               with parsed JSON (:func:`json.loads` by
                               default).
@@ -894,7 +894,7 @@ WebSocketResponse
 
    .. coroutinemethod:: receive_str()
 
-      A :ref:`coroutine<coroutine>` that calls :meth:`receive_mgs` but
+      A :ref:`coroutine<coroutine>` that calls :meth:`receive` but
       also asserts the message type is
       :const:`~aiohttp.websocket.MSG_TEXT`.
 
@@ -904,13 +904,29 @@ WebSocketResponse
 
    .. coroutinemethod:: receive_bytes()
 
-      A :ref:`coroutine<coroutine>` that calls :meth:`receive_mgs` but
+      A :ref:`coroutine<coroutine>` that calls :meth:`receive` but
       also asserts the message type is
       :const:`~aiohttp.websocket.MSG_BINARY`.
 
       :return bytes: peer's message content.
 
       :raise TypeError: if message is :const:`~aiohttp.websocket.MSG_TEXT`.
+
+   .. coroutinemethod:: recieve_json(*, loads=json.loads)
+
+      A :ref:`coroutine<coroutine>` that calls :meth:`receive`, asserts the
+      message type is :const:`~aiohttp.websocket.MSG_TEXT`, and loads the JSON
+      string to a Python dict.
+
+      :param callable loads: any :term:`callable` that accepts
+                              :class:`str` and returns :class:`dict`
+                              with parsed JSON (:func:`json.loads` by
+                              default).
+
+      :return dict: loaded JSON content
+
+      :raise TypeError: if message is :const:`~aiohttp.websocket.MSG_BINARY`.
+      :raise ValueError: if message is not valid JSON.
 
 
 .. versionadded:: 0.14
